backend/requirements.txt
fastapi
uvicorn
python-dotenv
duckduckgo-search
playwright
beautifulsoup4
google-generativeai
pydantic
